src/requirements.txt
albumentations==1.1.0
Cython==0.29.26
DCNv2.egg==info
easydict==1.9
ipython==8.3.0
matplotlib==3.3.4
numba==0.53.1
numpy==1.19.5
opencv_python_headless==4.5.5.62
pandas==1.1.5
Pillow==9.1.0
progress==1.6
pycocotools==2.0.4
PyYAML==6.0
scikit_spatial==5.2.0
scipy==1.5.4
setuptools==58.0.4
six==1.16.0
tensorboardX==2.5
torch==1.7.0
torchsampler==0.1.1
torchvision==0.8.0
